Refactor/reorganize/lint code
