+++ plugins/inputs/opensearch_query/aggregation.bucket_data.go
@@ -1,0 +1,1 @@
+package opensearch_query
